SpellChecker/requirements.txt
pyspellchecker==0.7.2
pytest==7.4.3